+++ frontend/features/page/CardListSectionBlock.tsx
