+++ frontend/features/page/FeatureListSectionBlock.tsx
