+++ frontend/features/page/HeroSectionBlock.tsx
